Add support to build everything in pixi (#414)
This builds on #409 to add pixi support for snax-cluster.

Previously, the build only supported to be built inside the container, this should provide an environment in which that is possible without container, which is nice for running on supercomputers, and for running in other complex environments which require different other packages.

---------

Co-authored-by: Joren Dumoulin <joren.dumoulin@kuleuven.be>
Co-authored-by: Josse Van Delm <josse.vandelm@esat.kuleuven.be>